Rename github/models to github/trees
